requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
python-dotenv==1.0.0